+++ internal/procgroup/session_windows.go
@@ -7,3 +7,5 @@ func guardedSessionSupport() error { return errUnsupported }
 func sessionIdentity(int) (int, string, error) { return 0, "", errUnsupported }
 
 func signalGuardSession(int, bool) error { return errUnsupported }
+
+func signalDiagnosticSession(int) error { return errUnsupported }